add client-side tests for event user properties
